PIM-5685: Move TranslatedLabelsProviderInterface from repository to provider directory
