+++ spec/Pim/Bundle/EnrichBundle/Doctrine/ORM/Repository/AttributeRepositorySpec.php
@@ -30,7 +30,7 @@ class AttributeRepositorySpec extends ObjectBehavior
 
     function it_provides_translated_data()
     {
-        $this->shouldImplement('Pim\Component\Enrich\Repository\TranslatedLabelsProviderInterface');
+        $this->shouldImplement('Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface');
     }
 
     function it_is_a_doctrine_repository()
