+++ spec/Pim/Bundle/EnrichBundle/Doctrine/ORM/Repository/GroupTypeRepositorySpec.php
@@ -29,7 +29,7 @@ class GroupTypeRepositorySpec extends ObjectBehavior
 
     function it_provides_translated_data()
     {
-        $this->shouldImplement('Pim\Component\Enrich\Repository\TranslatedLabelsProviderInterface');
+        $this->shouldImplement('Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface');
     }
 
     function it_is_a_doctrine_repository()
