+++ src/Pim/Bundle/EnrichBundle/Doctrine/ORM/Repository/GroupTypeRepository.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\EnrichBundle\Doctrine\ORM\Repository;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityRepository;
 use Pim\Bundle\UserBundle\Context\UserContext;
-use Pim\Component\Enrich\Repository\TranslatedLabelsProviderInterface;
+use Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface;
 
 /**
  * @author    Arnaud Langlade <arnaud.langlade@akeneo.com>
