+++ src/Pim/Bundle/EnrichBundle/Form/Type/ChannelType.php
@@ -8,7 +8,7 @@ use Pim\Bundle\EnrichBundle\Helper\SortHelper;
 use Pim\Bundle\EnrichBundle\Provider\ColorsProvider;
 use Pim\Component\Catalog\Repository\CurrencyRepositoryInterface;
 use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
-use Pim\Component\Enrich\Repository\TranslatedLabelsProviderInterface;
+use Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\Form\FormInterface;
@@ -48,7 +48,7 @@ class ChannelType extends AbstractType
      * @param LocaleRepositoryInterface $localeRepository
      * @param LocaleHelper              $localeHelper
      * @param ColorsProvider            $provider
-     * @param TranslatedLabelsProviderInterface  $categoryRepository
+     * @param \Pim\Component\Enrich\Provider\TranslatedLabelsProviderInterface  $categoryRepository
      * @param string                    $categoryClass
      * @param string                    $dataClass
      */
